low ap items

+++ yolo/tasks/yolo.py
@@ -352,7 +352,7 @@ class YoloTask(base_task.Task):
     #tf.print(optimizer.iterations, loss_metrics['global']['total_loss'])
 
     logs = {self.loss: loss_metrics['global']['total_loss'],
-            "iteration": optimizer.iterations}
+            "iteration": tf.cast(optimizer.iterations, tf.float32)}
     if metrics:
       for m in metrics:
         m.update_state(loss_metrics[m.name])
